When called with -workspaces, do not map initial frame.
The initial frame is temporary and used to be created as a normal frame
which was then deleted.  But after change 200938b95d (Po Lu
<luangruo@yahoo.com> 2022-07-15), emacs will attempt to run
move-frame-functions hooks afterward, on what is now a dead frame.

Whether or not that's a bug, we can probably avoid a lot of setup overhead
and future interaction bugs by never mapping the temporary frame in the
first place.